+++ tests/math.c
@@ -138,7 +138,7 @@ int main(int argc, char* argv[])
     const unsigned int test[] = { 0, 1, 2, 3, 5, 7, 12, 13, 24, 32, 2057, 120, 14, 997 };
     const int n = sizeof(test) / sizeof(*test);
     for (i = 0; i < n; ++i) {
-      const unsigned int coprime = libxsmm_shuffle(test[i]);
+      const size_t coprime = libxsmm_shuffle(test[i]);
       const unsigned int gcd = (unsigned int)libxsmm_gcd(coprime, test[i]);
       if ((0 != coprime || 1 < test[i]) && (test[i] <= coprime || 1 != gcd)) {
         exit(EXIT_FAILURE);
